Tweak visuals
* add camera rotation
* add number key selecting
* replace planes with models
* replace phong with blinn
* fix animation system
* precalculate object matrices
* remove unnecessary light properties
* show enemy tiles only when match starts

+++ src/utils/widgets.cpp
@@ -493,7 +493,7 @@ void LabelEdit::setText(const string& str) {
 
 void LabelEdit::onTextReset() {
 	updateTextTex();
-	cpos = text.length();
+	cpos = uint(text.length());
 	textOfs = 0;
 }
